added failing tests for multiple removeFile
I added failing tests to prove it's failing before I submit the fix.
When calling removeFile many times some files are not removed. The use
case is add 2 files and have a button to remove each. These buttons call
removeFile. At the end of the operations we have 1 file listed, instead
of zero.