fix multiple removeFile calls
The issue was caused by outdated "files" variable. This is due to how
closure works. Each removeFile starts with 2 files in "files". Then when
we remove the file, one file remains. Then when the second removeFile
file is called, one file remains. The solution is using functional
updates, which makes sure we are calculating state over the previous one
instead of some outdated value due to closure.

Reference:

1. https://reactjs.org/docs/hooks-reference.html#functional-updates